reorganized structure for better testability, added some tests

+++ src/Main.cpp
@@ -35,13 +35,13 @@ bool Main::OnInit()
 	#ifdef _DEBUG
 		#ifdef _WIN32
 			int hConHandle;
-			long lStdHandle;
+			long long lStdHandle;
 			FILE *fp;
 			// allocate a console for this app
 			AllocConsole();
 
 			// redirect unbuffered STDOUT to the console
-			lStdHandle = (long)GetStdHandle(STD_OUTPUT_HANDLE);
+			lStdHandle = (long long)GetStdHandle(STD_OUTPUT_HANDLE);
 			hConHandle = _open_osfhandle(lStdHandle, _O_TEXT);
 			fp = _fdopen(hConHandle, "w");
 			*stdout = *fp;
